Renamed jwt helpers folder name.
