+++ backend/utils/jwtHelpers/destroyAuthToken.js
