+++ backend/utils/jwtHelpers/generateAuthToken.js
